+++ pkg/codegen/hcl2/syntax/utilities.go
@@ -1,0 +1,6 @@
+package syntax
+
+import "github.com/hashicorp/hcl/v2/hclsyntax"
+
+// None is an HCL syntax node that can be used when a syntax node is required but none is appropriate.
+var None hclsyntax.Node = &hclsyntax.Body{}